Remove redundant casts

+++ src/engine/client/graphics_threaded.cpp
@@ -2022,7 +2022,7 @@ void *CGraphics_Threaded::AllocCommandBufferData(size_t AllocSize)
 		if(pData == nullptr)
 		{
 			char aError[256];
-			str_format(aError, sizeof(aError), "graphics: failed to allocate data (size %" PRIzu ") for command buffer", (size_t)AllocSize);
+			str_format(aError, sizeof(aError), "graphics: failed to allocate data (size %" PRIzu ") for command buffer", AllocSize);
 			dbg_assert(false, aError);
 		}
 	}
